# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SafePlayer API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /media:
    get:
      operationId: listMedia
      tags: [media]
      summary: List recently viewed media for current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MediaEntry"

    post:
      operationId: createMedia
      tags: [media]
      summary: Add media to library (file or URL)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MediaInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaEntry"
  /media/{id}:
    get:
      operationId: getMedia
      tags: [media]
      summary: Get media entry with jump frames and safety status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaDetail"
        "404":
          description: Not found

    delete:
      operationId: deleteMedia
      tags: [media]
      summary: Remove media from library
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /media/lookup:
    post:
      operationId: lookupMedia
      tags: [media]
      summary: Lookup safety status by title or file hash
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MediaLookupInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaLookupResult"

  # JUMP FRAMES
...
  /submissions:
    get:
      operationId: listSubmissions
      tags: [submissions]
      summary: List pending community submissions (admin only)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, approved, rejected, flagged]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Submission"
  /submissions/{id}/approve:
    post:
      operationId: approveSubmission
      tags: [submissions]
      summary: Admin approves a submission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
  /submissions/{id}/reject:
    post:
      operationId: rejectSubmission
      tags: [submissions]
      summary: Admin rejects a submission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"

  # ADMIN
...
components:
  schemas:
...
    MediaInput:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [file, url]
        title:
          type: string
        fileName:
          type: ["string", "null"]
        fileHash:
          type: ["string", "null"]
        url:
          type: ["string", "null"]
        mimeType:
          type: ["string", "null"]
    MediaEntry:
      type: object
      required: [id, type, title, safetyStatus, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [file, url]
        title:
          type: string
        fileName:
          type: ["string", "null"]
        fileHash:
          type: ["string", "null"]
        url:
          type: ["string", "null"]
        mimeType:
          type: ["string", "null"]
        safetyStatus:
          type: string
          enum: [safe, unpreviewed, flagged]
        jumpFrameCount:
          type: integer
        lastWatched:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    MediaDetail:
      type: object
      required: [media, jumpFrames, safetyStatus]
      properties:
        media:
          $ref: "#/components/schemas/MediaEntry"
        jumpFrames:
          type: array
          items:
            $ref: "#/components/schemas/JumpFrame"
        safetyStatus:
          type: string
          enum: [safe, unpreviewed, flagged]
        aiFlag:
          $ref: "#/components/schemas/AiFlag"
    MediaLookupInput:
      type: object
      properties:
        title:
          type: ["string", "null"]
        fileHash:
          type: ["string", "null"]
    MediaLookupResult:
      type: object
      required: [found, safetyStatus]
      properties:
        found:
          type: boolean
        mediaId:
          type: ["integer", "null"]
        safetyStatus:
          type: string
          enum: [safe, unpreviewed, flagged]
        jumpFrameCount:
          type: integer
        source:
          type: ["string", "null"]
          enum: [personal, global, null]

    # JUMP FRAME SCHEMAS
    JumpFrame:
      type: object
      required: [id, mediaId, startTime, endTime, category, source, validated, createdAt]
      properties:
        id:
          type: integer
        mediaId:
          type: integer
        startTime:
          type: number
          description: seconds
        endTime:
          type: number
          description: seconds
        category:
          type: string
          enum: [violence, sexual, language, other]
        source:
          type: string
          enum: [personal, global]
        validated:
          type: boolean
        submittedBy:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
...
    Submission:
      type: object
      required: [id, mediaId, startTime, endTime, category, status, submittedBy, createdAt]
      properties:
        id:
          type: integer
        mediaId:
          type: integer
        mediaTitle:
          type: ["string", "null"]
        startTime:
          type: number
        endTime:
          type: number
        category:
          type: string
          enum: [violence, sexual, language, other]
        status:
          type: string
          enum: [pending, approved, rejected, flagged]
        submittedBy:
          type: integer
        submitterEmail:
          type: ["string", "null"]
        aiFlag:
          $ref: "#/components/schemas/AiFlag"
        createdAt:
          type: string
          format: date-time
    AiFlag:
      type: ["object", "null"]
      properties:
        flagged:
          type: boolean
        reason:
          type: ["string", "null"]
        severity:
          type: string
          enum: [low, medium, high, extreme]
...