# lib/api-zod/src/generated/types/mediaLookupInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SafePlayer API
 * OpenAPI spec version: 0.1.0
 */

export interface MediaLookupInput {
  /** @nullable */
  title?: string | null;
  /** @nullable */
  fileHash?: string | null;
}