# lib/api-zod/src/generated/types/mediaLookupResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SafePlayer API
 * OpenAPI spec version: 0.1.0
 */
import type { MediaLookupResultSafetyStatus } from './mediaLookupResultSafetyStatus';
import type { MediaLookupResultSource } from './mediaLookupResultSource';

export interface MediaLookupResult {
  found: boolean;
  /** @nullable */
  mediaId?: number | null;
  safetyStatus: MediaLookupResultSafetyStatus;
  jumpFrameCount?: number;
  /** @nullable */
  source?: MediaLookupResultSource;
}